+++ src/configuration/ConfigEditor.tsx
@@ -21,13 +21,14 @@ import React, { useRef } from 'react';
 import { SIGV4ConnectionConfig } from '@grafana/aws-sdk';
 import { DataSourcePluginOptionsEditorProps, DataSourceSettings } from '@grafana/data';
 import { config } from '@grafana/runtime';
-import { AlertingSettings, DataSourceHttpSettings, Alert, Icon } from '@grafana/ui';
+import { AlertingSettings, DataSourceHttpSettings } from '@grafana/ui';
 
 import { PromOptions } from '../types';
 
 import { AzureAuthSettings } from './AzureAuthSettings';
 import { hasCredentials, setDefaultCredentials, resetCredentials } from './AzureCredentialsConfig';
 import { PromSettings } from './PromSettings';
+import { TipsSetup } from "./TipsSetup";
 
 export enum DataSourceType {
   Alertmanager = 'alertmanager',
@@ -50,48 +51,17 @@ export const ConfigEditor = (props: Props) => {
 
   return (
     <>
-      {options.access === 'direct' && (
-        <Alert title="Error" severity="error">
-          Browser access mode in the Prometheus datasource is no longer available. Switch to server access mode.
-        </Alert>
-      )}
+      <TipsSetup/>
 
-      <div className="gf-form-group">
-        <h3 className="page-heading">Tips on how to setup</h3>
-
-       <div className="submenu-controls">
-         <div className="gf-form">
-           <a
-             className="gf-form-label gf-form-label--dashlink"
-             href="https://docs.victoriametrics.com/Cluster-VictoriaMetrics.html#url-format"
-             target="_blank"
-             rel="noreferrer">
-             <Icon name="file-alt"/>
-             <span style={{ marginLeft: "4px" }}>Cluster VM</span>
-           </a>
-         </div>
-         <div className="gf-form">
-           <a
-             className="gf-form-label gf-form-label--dashlink"
-             href="https://docs.victoriametrics.com/Single-server-VictoriaMetrics.html#grafana-setup"
-             target="_blank"
-             rel="noreferrer">
-             <Icon name="file-alt"/>
-             <span style={{ marginLeft: "4px" }}>Grafana setup</span>
-           </a>
-         </div>
-       </div>
-      </div>
-
-        <DataSourceHttpSettings
-          defaultUrl="http://localhost:8428"
-          dataSourceConfig={options}
-          showAccessOptions={showAccessOptions.current}
-          onChange={onOptionsChange}
-          sigV4AuthToggleEnabled={config.sigV4AuthEnabled}
-          azureAuthSettings={azureAuthSettings}
-          renderSigV4Editor={<SIGV4ConnectionConfig {...props}></SIGV4ConnectionConfig>}
-        />
+      <DataSourceHttpSettings
+        defaultUrl="http://localhost:8428"
+        dataSourceConfig={options}
+        showAccessOptions={showAccessOptions.current}
+        onChange={onOptionsChange}
+        sigV4AuthToggleEnabled={config.sigV4AuthEnabled}
+        azureAuthSettings={azureAuthSettings}
+        renderSigV4Editor={<SIGV4ConnectionConfig {...props}></SIGV4ConnectionConfig>}
+      />
 
       <AlertingSettings<PromOptions>
         alertmanagerDataSources={alertmanagers}
@@ -99,7 +69,7 @@ export const ConfigEditor = (props: Props) => {
         onOptionsChange={onOptionsChange}
       />
 
-      <PromSettings options={options} onOptionsChange={onOptionsChange} />
+      <PromSettings options={options} onOptionsChange={onOptionsChange}/>
     </>
   );
 };